refactor(node-filter): extract node filter
